+++ app/src/main/java/com/sg2022/we_got_the_moves/repository/WorkoutsRepository.java
@@ -82,7 +82,7 @@ public class WorkoutsRepository {
     return this.workoutDao.getWorkout(id);
   }
 
-  public LiveData<List<Exercise>> getAllExercises(int workoutId) {
+  public LiveData<List<Exercise>> getAllExercises(long workoutId) {
     return this.exerciseDao.getAllExercises(workoutId);
   }
